src: arreglos en el repositorio de persona humana

+++ src/main/java/ar/edu/utn/frba/dds/domain/repositories/imp/RepositorioPersonaHumana.java
@@ -16,15 +16,15 @@ public class RepositorioPersonaHumana implements IRepositorioPersonaHumana {
 
   @Override
   public Long guardar(PersonaHumana persona) {
-    persona.setId(this.personas.size() + 1);
+    persona.setId((long) (this.personas.size() + 1));
     this.personas.add(persona);
     return persona.getId();
   }
 
   @Override
   public void actualizar(PersonaHumana persona) {
-    for(int i=0; i < personas.size(); i++){
-      if(persona.getId().equals(personas.get(i).getId())){
+    for (int i = 0; i < personas.size(); i++) {
+      if (persona.getId().equals(personas.get(i).getId())) {
         personas.set(i, persona);
         break;
       }
